Delete prefix module, move get_executable_filename
