--- libs/pika/prefix/README.rst
@@ -1,13 +1,0 @@
-
-..
-    Copyright (c) 2019 The STE||AR-Group
-
-    SPDX-License-Identifier: BSL-1.0
-    Distributed under the Boost Software License, Version 1.0. (See accompanying
-    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-======
-prefix
-======
-
-This library is part of pika.
