--- libs/pika/prefix/docs/index.rst
@@ -1,18 +1,0 @@
-..
-    Copyright (c) 2019 The STE||AR-Group
-
-    SPDX-License-Identifier: BSL-1.0
-    Distributed under the Boost Software License, Version 1.0. (See accompanying
-    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-.. _modules_prefix:
-
-======
-prefix
-======
-
-This module provides utilities for handling the prefix of an |pika| application,
-i.e. the paths used for searching components and plugins.
-
-See the :ref:`API reference <modules_prefix_api>` of this module for more
-details.
